requirements.txt
ansible-runner==2.2.0
boto3==1.23.8
botocore==1.26.8
decorator==5.1.1
docutils==0.18.1
jmespath==0.10.0
jsonpath-ng==1.5.3
lockfile==0.12.2
packaging==21.3
pexpect==4.8.0
ply==3.11
ptyprocess==0.7.0
pyparsing==3.0.9
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
s3transfer==0.5.2
six==1.16.0
structlog==21.5.0
typing_extensions==4.1.1
urllib3==1.26.9
